app/requirements.txt
django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-otp>=1.3.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.0
Pillow>=10.2.0
django-storages>=1.14.2
google-cloud-storage>=2.14.0
django-ratelimit>=4.1.0
django-axes>=6.3.0
django-ipware>=6.0.4
django-user-agents>=0.4.0
django-request-logging>=0.7.0
django-session-security>=2.8.0
django-password-validators>=1.3.0
django-sslserver>=0.22
django-honeypot>=1.0.2
django-defender>=0.8.4 